More modernization.

+++ make/src/FSItemIndex.m
@@ -125,11 +125,11 @@
 
 - (void) addItemsFromArray: (NSArray*) items
 {
-	unsigned int i = [items count];
+	NSUInteger i = [items count];
 	
 	NSAutoreleasePool *localPool = (i > 200) ? [[NSAutoreleasePool alloc] init] : nil;
 	
-	unsigned poolLoopCount = 0;
+	NSUInteger poolLoopCount = 0;
 	
 	while ( i-- )
 	{
