+++ make/src/NSOutlineView-ContextMenuExtension.m
@@ -12,7 +12,7 @@
 // return the selected item
 - (id) selectedItem
 {
-    int row = [self selectedRow];
+    NSInteger row = [self selectedRow];
     return row >= 0 ? [self itemAtRow: row] : nil;
 }
 
@@ -21,8 +21,8 @@
 {
     NSPoint point = [self convertPoint: [evt locationInWindow] fromView: nil];
     
-    int columnIndex = [self columnAtPoint:point];
-    int rowIndex = [self rowAtPoint:point];
+    NSInteger columnIndex = [self columnAtPoint:point];
+    NSInteger rowIndex = [self rowAtPoint:point];
 	
     if ( rowIndex >= 0 && [self numberOfSelectedRows] <= 1)
         [self selectRow: rowIndex byExtendingSelection: NO];
